# perf(ecc): faster affine Add