… ecc/bls12-377/g1_test.go
… ecc/bls12-377/g2_test.go
… ecc/bls12-378/g1_test.go
… ecc/bls12-378/g2_test.go
… ecc/bls12-381/g1_test.go
… ecc/bls12-381/g2_test.go
… ecc/bls24-315/g1_test.go
… ecc/bls24-315/g2_test.go
… ecc/bls24-317/g1_test.go
… ecc/bls24-317/g2_test.go
… ecc/bn254/g1_test.go
… ecc/bn254/g2_test.go
… ecc/bw6-633/g1_test.go
… ecc/bw6-633/g2_test.go
… ecc/bw6-756/g1_test.go
… ecc/bw6-756/g2_test.go
… ecc/bw6-761/g1_test.go
… ecc/bw6-761/g2_test.go
… ecc/secp256k1/g1.go
… ecc/secp256k1/g1_test.go
… internal/generator/ecc/template/point.go.tmpl
… internal/generator/ecc/template/tests/point.go.tmpl