# Description

Actually, we can optimize affine addition more than #509 with the assumption that both `Z1` and `Z2` are 1 in Jacobian coordinates.  This PR uses [these EFD formulas](https://www.hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#addition-mmadd-2007-bl). Also #509 modifies the second input in the `Sub` method which this PR fixes. 
https://github.com/Consensys/gnark-crypto/blob/364bbceb6d746d027c0e3f78912e61a129b51b8c/internal/generator/ecc/template/point.go.tmpl#L128

## Type of change

- [x] New feature (non-breaking change which adds functionality)

# How has this been tested?

same tests as #509 

# How has this been benchmarked?

e.g. BLS12-381 G1 addition on a z1d.large AWS machine:
```
goos: linux
goarch: amd64
pkg: github.com/consensys/gnark-crypto/ecc/bls12-381
cpu: Intel(R) Xeon(R) Platinum 8151 CPU @ 3.40GHz
BenchmarkG1AffineAdd
BenchmarkG1AffineAdd-2   	  232734	      5135 ns/op
PASS
```
This is a -3.56% saving compared to #509 
```
benchmark                  old ns/op     new ns/op     delta
BenchmarkG1AffineAdd-2     5135          4952          -3.56%
```

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I did not modify files generated from templates
- [x] `golangci-lint` does not output errors locally
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules