@yelhousni 2024-06-12
> Code looks correct and match the formulas, however, it is not tested (i.e. G1Affine.Add is not called in any test files, only in `fr/lookup` and `kzg` ).
> 
> Can you add some basic tests / property based tests now that we have a different code path than `Jacobian`?
> 
> (general addition, doubling, adding of point at infinity, and addition resulting in point at infinity)

I think in `TestG*AffineOps` we call affine `Add`, `Sub`, `Double` and `ScalarMul` here for example:
https://github.com/Consensys/gnark-crypto/blob/3ba464d9dfe41bb88458da9f687d78cf9a5c4838/ecc/bls12-381/g1_test.go#L266

This test
https://github.com/Consensys/gnark-crypto/blob/3ba464d9dfe41bb88458da9f687d78cf9a5c4838/ecc/bls12-381/g1_test.go#L250
checks the edge case where inputs are equal in `Add` but I do agree that tests involving the point `(0,0)` are missing. Will add.